Correccionde modelos y migraciones. Creacion de vista departamento. Implementacion de componentes preline. Implementacion de validacion de campos en empleados. Creacion de vista Retardos y manejo de importacion de datos desde archivos excel.

+++ database/migrations/2025_07_01_213021_create_prestamo_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->date('fecha_fin');
             $table->unsignedInteger('pago_periodo');
             $table->unsignedInteger('total');
-            $table->unsignedInteger('id_estatus');
+            $table->unsignedBigInteger('id_estatus');
             $table->timestamps();
         });
